Update RubyGems to the latest release all the time
RubyGems 2.7.3 (the one currently used in building 2.5.x) has a bug
that affects `$LOAD_PATH`. This manifests in mysterious failures of
`bundle exec rake`, where `bundler/dep_proxy` (or other files) are
not found.

We avoid this bug by updating RubyGems to at least 2.7.7 (the current
release as of this writing).